Optimization ideas (#1)

+++ contracts/ILifeOnchain.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.17;
  * @title Traits indexes for each life
  */
 struct Traits {
-    uint256 speedIndex;
-    uint256 colorIndex;
-    uint256 modeIndex;
+    uint16 speedIndex;
+    uint16 colorIndex;
+    uint16 modeIndex;
 }
